Improve ANSI color readability in report error details
