+++ packages/web-awesome/src/components/TestResult/TrError/index.tsx
@@ -1,5 +1,5 @@
 import type { TestError, TestStatus } from "@allurereport/core-api";
-import { ansiToHTML } from "@allurereport/web-commons";
+import { ansiSemanticColors, ansiToHTML, normalizeAnsiForegroundColors } from "@allurereport/web-commons";
 import { Button, Code, IconButton, Text, TooltipWrapper, allureIcons } from "@allurereport/web-components";
 import clsx from "clsx";
 import { type FunctionalComponent } from "preact";
@@ -13,19 +13,15 @@ import { copyToClipboard } from "@/utils/copyToClipboard";
 
 import * as styles from "./styles.scss";
 
-const TrErrorTrace = ({ trace }: { trace: string }) => {
-  const sanitizedTrace = ansiToHTML(trace, {
+const ansiStatusDetailsToHTML = (text: string) =>
+  ansiToHTML(normalizeAnsiForegroundColors(text), {
     fg: "var(--color-text-primary)",
     bg: "none",
-    colors: {
-      0: "none",
-      1: "none",
-      2: "var(--color-intent-info-text)",
-      3: "var(--color-intent-warning-text)",
-      4: "var(--color-status-unknown-chart)",
-      5: "var(--color-syntax-type)",
-    },
+    colors: ansiSemanticColors,
   });
+
+const TrErrorTrace = ({ trace }: { trace: string }) => {
+  const sanitizedTrace = ansiStatusDetailsToHTML(trace);
 
   return (
     <div data-testid="test-result-error-trace" className={styles["test-result-error-trace"]}>
@@ -52,13 +48,7 @@ export const TrError: FunctionalComponent<
       data: { actual, expected },
       component: <TrDiff actual={actual} expected={expected} />,
     });
-  const sanitizedMessage =
-    showMessage && message
-      ? ansiToHTML(message, {
-          fg: "var(--color-text-primary)",
-          colors: {},
-        })
-      : "";
+  const sanitizedMessage = showMessage && message ? ansiStatusDetailsToHTML(message) : "";
 
   return (
     <div
